requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
dinjo==0.1.0
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
mongoengine==0.23.0
mongomock==3.22.1
numpy==1.20.3
packaging==20.9
pandas==1.2.4
plotly==4.14.3
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pydantic==1.8.1
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2021.1
requests==2.25.1
retrying==1.3.3
rsa==4.7.2
scipy==1.6.3
sentinels==1.0.0
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.4
uvicorn==0.13.4
